requirements.txt
python-whois
dnspython
requests
rich
beautifulsoup4
pytest
